Update open api script
