+++ scripts/open-api/public/index.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet" />
+    <script type="module" src="https://unpkg.com/rapidoc/dist/rapidoc-min.js"></script>
+  </head>
+  <body>
+    <rapi-doc
+      spec-url="/spec.yml"
+      allow-server-selection="false"
+      show-header="false"
+      regular-font="Nunito"
+      bg-color="#2B303B"
+      text-color="#dee3ec"
+      nav-bg-color=""
+      nav-text-color=""
+      nav-hover-bg-color=""
+      nav-hover-text-color=""
+      nav-accent-color=""
+      primary-color=""
+      theme="dark"
+
+      schema-style="table"
+    >
+    </rapi-doc>
+  </body>
+</html>
